Rename core/rapid to index. Update paths associated
